[Fixes #8689] Remove endpoint from each resorce, keep it only on base resource
