+++ geonode/base/api/tests.py
@@ -36,12 +36,13 @@ from guardian.shortcuts import get_anonymous_user
 
 from geonode import geoserver
 from geonode.layers.models import Layer
+from geonode.tests.base import GeoNodeBaseTestSupport
 from geonode.utils import check_ogc_backend, set_resource_default_links
 from geonode.favorite.models import Favorite
 from geonode.documents.models import Document
 from geonode.base.utils import build_absolute_uri
 from geonode.thumbs.exceptions import ThumbnailError
-from geonode.base.populate_test_data import create_models
+from geonode.base.populate_test_data import create_models, create_single_layer
 from geonode.security.utils import get_resources_with_perms
 
 from geonode.base.models import (
@@ -1021,3 +1022,61 @@ class BaseApiTests(APITestCase, URLPatternsTestCase):
         }
         self.assertEqual(response.status_code, 500)
         self.assertEqual(expected, response.json())
+
+
+class TestExtraMetadataLayersApi(GeoNodeBaseTestSupport):
+    def setUp(self):
+        self.layer = create_single_layer('single_layer')
+        self.metadata = {
+            "name": "metadata-name",
+            "slug": "metadata-slug",
+            "help_text": "this is the help text",
+            "field_type": "str",
+            "value": "my value",
+            "category": "cat1"
+        }
+        Layer.objects.filter(id=self.layer.id).update(extra_metadata=[self.metadata])
+
+    def test_get_will_return_the_list_of_extra_metadata(self):
+        self.client.login(username="admin", password="admin")
+        url = reverse('resources-extra-metadata', args=[self.layer.id])
+        response = self.client.get(url, content_type='application/json')
+        self.assertTrue(200, response.status_code)
+        self.assertEqual([self.metadata], response.json())
+
+    def test_put_will_update_the_whole_metadata(self):
+        self.client.login(username="admin", password="admin")
+        url = reverse('resources-extra-metadata', args=[self.layer.id])
+        input_metadata = {
+            "name": "metadata-updated",
+            "slug": "metadata-slug-updated",
+            "help_text": "this is the help text-updated",
+            "field_type": "str-updated",
+            "value": "my value-updated",
+            "category": "cat1-updated"
+        }
+        response = self.client.put(url, data=[input_metadata], content_type='application/json')
+        self.assertTrue(200, response.status_code)
+        self.assertEqual([input_metadata], response.json())
+
+    def test_post_will_add_new_metadata(self):
+        self.client.login(username="admin", password="admin")
+        url = reverse('resources-extra-metadata', args=[self.layer.id])
+        input_metadata = {
+            "name": "metadata-new",
+            "slug": "metadata-slug-new",
+            "help_text": "this is the help text-new",
+            "field_type": "str-new",
+            "value": "my value-new",
+            "category": "cat1-new"
+        }
+        response = self.client.post(url, data=[input_metadata], content_type='application/json')
+        self.assertTrue(201, response.status_code)
+        self.assertEqual(2, len(response.json()))
+
+    def test_delete_will_delete_single_metadata(self):
+        self.client.login(username="admin", password="admin")
+        url = reverse('resources-extra-metadata', args=[self.layer.id])
+        response = self.client.delete(url, data=[self.metadata], content_type='application/json')
+        self.assertTrue(200, response.status_code)
+        self.assertEqual([], response.json())
