+++ geonode/layers/api/tests.py
@@ -24,9 +24,8 @@ from rest_framework.test import APITestCase, URLPatternsTestCase
 
 from geonode import geoserver
 from geonode.layers.models import Layer
-from geonode.tests.base import GeoNodeBaseTestSupport
 from geonode.utils import check_ogc_backend
-from geonode.base.populate_test_data import create_models, create_single_layer
+from geonode.base.populate_test_data import create_models
 
 logger = logging.getLogger(__name__)
 
@@ -97,61 +96,3 @@ class LayersApiTests(APITestCase, URLPatternsTestCase):
         self.assertEqual(response.data['layer']['raw_constraints_other'], "None")
         self.assertEqual(response.data['layer']['raw_supplemental_information'], "No information provided í £682m")
         self.assertEqual(response.data['layer']['raw_data_quality_statement'], "OK    1 2   a b")
-
-
-class TestExtraMetadataLayersApi(GeoNodeBaseTestSupport):
-    def setUp(self):
-        self.layer = create_single_layer('single_layer')
-        self.metadata = {
-            "name": "metadata-name",
-            "slug": "metadata-slug",
-            "help_text": "this is the help text",
-            "field_type": "str",
-            "value": "my value",
-            "category": "cat1"
-        }
-        Layer.objects.filter(id=self.layer.id).update(extra_metadata=[self.metadata])
-
-    def test_get_will_return_the_list_of_extra_metadata(self):
-        self.client.login(username="admin", password="admin")
-        url = reverse('layers-extra-metadata', args=[self.layer.id])
-        response = self.client.get(url, content_type='application/json')
-        self.assertTrue(200, response.status_code)
-        self.assertEqual([self.metadata], response.json())
-
-    def test_put_will_update_the_whole_metadata(self):
-        self.client.login(username="admin", password="admin")
-        url = reverse('layers-extra-metadata', args=[self.layer.id])
-        input_metadata = {
-            "name": "metadata-updated",
-            "slug": "metadata-slug-updated",
-            "help_text": "this is the help text-updated",
-            "field_type": "str-updated",
-            "value": "my value-updated",
-            "category": "cat1-updated"
-        }
-        response = self.client.put(url, data=[input_metadata], content_type='application/json')
-        self.assertTrue(200, response.status_code)
-        self.assertEqual([input_metadata], response.json())
-
-    def test_post_will_add_new_metadata(self):
-        self.client.login(username="admin", password="admin")
-        url = reverse('layers-extra-metadata', args=[self.layer.id])
-        input_metadata = {
-            "name": "metadata-new",
-            "slug": "metadata-slug-new",
-            "help_text": "this is the help text-new",
-            "field_type": "str-new",
-            "value": "my value-new",
-            "category": "cat1-new"
-        }
-        response = self.client.post(url, data=[input_metadata], content_type='application/json')
-        self.assertTrue(201, response.status_code)
-        self.assertEqual(2, len(response.json()))
-
-    def test_delete_will_delete_single_metadata(self):
-        self.client.login(username="admin", password="admin")
-        url = reverse('layers-extra-metadata', args=[self.layer.id])
-        response = self.client.delete(url, data=[self.metadata], content_type='application/json')
-        self.assertTrue(200, response.status_code)
-        self.assertEqual([], response.json())
