+++ geonode/layers/api/views.py
@@ -22,11 +22,8 @@ from dynamic_rest.filters import DynamicFilterBackend, DynamicSortingFilter
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.authentication import SessionAuthentication, BasicAuthentication
 from oauth2_provider.contrib.rest_framework import OAuth2Authentication
-from drf_spectacular.utils import extend_schema
-from rest_framework.decorators import action
-from geonode.base.api.views import common_extra_metadata_handler
 from geonode.base.api.filters import DynamicSearchFilter, ExtentFilter
-from geonode.base.api.permissions import IsOwnerOrAdmin, IsOwnerOrReadOnly
+from geonode.base.api.permissions import IsOwnerOrReadOnly
 from geonode.base.api.pagination import GeoNodeApiPagination
 from geonode.layers.models import Layer
 
@@ -51,18 +48,3 @@ class LayerViewSet(DynamicModelViewSet):
     queryset = Layer.objects.all()
     serializer_class = LayerSerializer
     pagination_class = GeoNodeApiPagination
-
-    @extend_schema(
-        methods=["get", "put", "delete", "post"], description="Get or update extra metadata for each resource"
-    )
-    @action(
-        detail=True,
-        methods=["get", "put", "delete", "post"],
-        permission_classes=[
-            IsOwnerOrAdmin,
-        ],
-        url_path=r"extra_metadata",  # noqa
-        url_name="extra-metadata",
-    )
-    def extra_metadata(self, request, pk=None):
-        return common_extra_metadata_handler(request, self.get_object(), self.queryset)
